Working Stage5

+++ src/main/java/com/recipes/presentation/RecipeController.java
@@ -5,6 +5,8 @@ import com.recipes.businesslayer.RecipeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import javax.validation.Valid;
@@ -47,11 +49,14 @@ public class RecipeController {
     }
 
     @PostMapping("/api/recipe/new")
-    public ResponseEntity<String> addRecipe(@Valid @RequestBody Recipe recipe) {
+    public ResponseEntity<String> addRecipe(@AuthenticationPrincipal UserDetails userDetails,
+                                            @Valid @RequestBody Recipe recipe) {
+        String name = userDetails.getUsername();
+
         try {
-            Recipe createdRecipe = recipeService.save(new Recipe(recipe.getId(), recipe.getName(), recipe.getCategory(),
+            Recipe createdRecipe = recipeService.save(new Recipe(recipe.getId(), recipe.getName(),recipe.getCategory(),
                                                       LocalDateTime.now(), recipe.getDescription(),
-                                                      recipe.getIngredients(), recipe.getDirections()));
+                                                      recipe.getIngredients(), recipe.getDirections(), name));
             return new ResponseEntity<>("{\"id\": " + createdRecipe.getId() + "}", HttpStatus.OK);
         } catch (RuntimeException e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
@@ -60,29 +65,42 @@ public class RecipeController {
 
     @DeleteMapping("/api/recipe/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteRecipe(@PathVariable long id) {
-        try {
-            recipeService.deleteRecipeById(id);
-        } catch (RuntimeException e) {
+    public void deleteRecipe(@AuthenticationPrincipal UserDetails userDetails, @PathVariable long id) {
+        if (recipeService.existsById(id)) {
+            if (recipeService.findRecipeById(id).getAuthor() == userDetails.getUsername()) {
+                try {
+                    recipeService.deleteRecipeById(id);
+                } catch (RuntimeException e) {
+                    throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+                }
+            } else {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
+        } else {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
     }
 
     @PutMapping("/api/recipe/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updateRecipe(@PathVariable long id, @Valid @RequestBody Recipe recipe) {
+    public void updateRecipe(@AuthenticationPrincipal UserDetails userDetails, @PathVariable long id,
+                             @Valid @RequestBody Recipe recipe) {
         if (recipeService.existsById(id)) {
-            try {
-                Recipe recipeToUpdate = recipeService.findRecipeById(id);
-                recipeToUpdate.setName(recipe.getName());
-                recipeToUpdate.setCategory(recipe.getCategory());
-                recipeToUpdate.setDate(LocalDateTime.now());
-                recipeToUpdate.setDescription(recipe.getDescription());
-                recipeToUpdate.setIngredients(recipe.getIngredients());
-                recipeToUpdate.setDirections(recipe.getDirections());
-                recipeService.save(recipeToUpdate);
-            } catch (RuntimeException e) {
-                throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            if (recipeService.findRecipeById(id).getAuthor() == (userDetails.getUsername())) {
+                try {
+                    Recipe recipeToUpdate = recipeService.findRecipeById(id);
+                    recipeToUpdate.setName(recipe.getName());
+                    recipeToUpdate.setCategory(recipe.getCategory());
+                    recipeToUpdate.setDate(LocalDateTime.now());
+                    recipeToUpdate.setDescription(recipe.getDescription());
+                    recipeToUpdate.setIngredients(recipe.getIngredients());
+                    recipeToUpdate.setDirections(recipe.getDirections());
+                    recipeService.save(recipeToUpdate);
+                } catch (RuntimeException e) {
+                    throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+                }
+            } else {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
             }
         } else {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
